Reorganize API functions.
